Introduce delays for files actively modified.
The idea is to way a bit (but not forever) if we notice that files are
actively worked on.

To do so cleanly I moved timing operations to the filestore, which is
now in charge of all decision making based on changes to file state.

This introduces the Timing struct which holds all the timing-related
information about scanning, hashing, etc.